refactor : kpiAtom, CardStatusData 파일명
